upgrade libsrtp to latest main; sync buildsystem
this changes from the stream-list-api-orig branch, which
was made on top of our current version of libsrtp, to the
stream-list-api branch in my fork, which is based on the
latest main. we thought this would be a nice opportunity
to upgrade and sync to upstream.

the .gyp adaptation is based on chromium, which AFAIK hasn't
upgraded upstream and are now based on GN anyway.

steps:

 - review Makefile.in diff and manually apply the relevant
   differences to the .gyp, which are mainly wrt file changes
 - sync the list of headers in the .gyp
 - copy include/srtp.h to include/srtp2/srtp.h
 - add GCM define which is required now
 - remove targets for the tests (I don't think it's worth it
   to make these out of date?)